PerceivedRisk shouldn't affect "completeness"
The role of perceived risk has been de-emphasized.  The existing code already did not use perceived risk as a factor in the total score, as it used the max score of the four factors.  The function to check if a risk score record was complete, however, the "perceived risk" as a required element (even though it is not).  This has been fixed.